redux store 구현
- store에 reducer 구현
